requirements.txt
asgiref==3.8.1
blinker==1.8.2
celery==5.3.6
click==8.1.7
colorama==0.4.6
dataclasses==0.6
distlib==0.3.9
Django>=4.2,<5.0
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==4.3.1
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dj-database-url==0.5.0
djongo==1.2.31
dnspython==2.7.0
filelock==3.18.0
Flask==2.2.2
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
numpy==1.25.1
pandas==2.2.2
platformdirs==4.3.7
psycopg2-binary==2.9.10
pymongo==4.12.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.1
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
virtualenv==20.30.0
Werkzeug==3.0.4

